+++ lib/src/command_type.dart
@@ -1,0 +1,1 @@
+class CommandType {}
