+++ lib/src/notice.dart
@@ -1,0 +1,1 @@
+class Notice {}
