+++ lib/src/query.dart
@@ -1,0 +1,1 @@
+class Query<T> {}
