+++ lib/src/row_description.dart
@@ -1,0 +1,1 @@
+class RowDescription {}
