+++ lib/src/uint_8_array.dart
@@ -1,0 +1,1 @@
+class Uint8Array {}
